+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * new config option `missingSteps` to setup different behavior when step definitions are missing ([#158](https://github.com/vitalets/playwright-bdd/issues/158))
 * new config option `matchKeywords` to enable keyword matching when searching for step definitions ([#221](https://github.com/vitalets/playwright-bdd/issues/221))
 * make config option `quote` default to `"single"` to have less escapes in the generated files
+* config option `featuresRoot` serves as a default directory for both `features` and `steps`, if these options are not explicitly defined
 * increase minimal Playwright version from 1.35 to 1.38
 
 ## 7.5.0